+++ packages/cbl/lib/src/install.dart
@@ -1,0 +1,15 @@
+export 'install/package.dart'
+    show
+        AndroidPackage,
+        Architecture,
+        ArchiveFormat,
+        DatabasePackageConfig,
+        Edition,
+        Library,
+        OS,
+        PackageConfig,
+        Package,
+        PackageLoader,
+        RemotePackageLoader,
+        StandardPackage;
+export 'install/utils.dart' show copyDirectoryContents, userCachesDir;