+++ packages/cbl_dart/lib/src/install_libraries.dart
@@ -1,17 +1,39 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:cbl/cbl.dart';
+// ignore: implementation_imports
+import 'package:cbl/src/install.dart';
+import 'package:crypto/crypto.dart';
 import 'package:path/path.dart' as p;
 
-import 'package.dart';
-import 'tools.dart';
-import 'utils.dart';
+import 'logging.dart';
+
+extension PackageMerging on Package {
+  static String signature(Iterable<Package> packages) {
+    final signatures =
+        packages.map((package) => package.signatureContent).toList()..sort();
+
+    return md5
+        .convert(utf8.encode(signatures.join()))
+        .bytes
+        .map((byte) => byte.toRadixString(16).padLeft(2, '0'))
+        .join();
+  }
+
+  String get signatureContent => [
+        config.library.name,
+        config.release,
+        if (config case DatabasePackageConfig(:final edition)) edition.name,
+        config.targetId
+      ].join();
+}
 
 Directory mergedNativeLibrariesInstallDir(
   Iterable<Package> packages,
   String directory,
 ) {
-  final signature = Package.mergedSignature(packages);
+  final signature = PackageMerging.signature(packages);
   return Directory(p.join(directory, signature));
 }
 
@@ -27,53 +49,16 @@ Future<void> installMergedNativeLibraries(
 }) async {
   logger.fine('Installing native libraries into $directory');
 
-  final tmpDir = await Directory.systemTemp.createTemp();
+  final installDir = mergedNativeLibrariesInstallDir(packages, directory);
+  await installDir.create(recursive: true);
 
-  try {
-    final tmpInstallDir = Directory.fromUri(tmpDir.uri.resolve('lib'));
-    await tmpInstallDir.create();
-
-    for (final package in packages) {
-      await installNativeLibrary(
-        package,
-        installDir: tmpInstallDir.path,
-        tmpDir: tmpDir.path,
-      );
-    }
-
-    final installDir = mergedNativeLibrariesInstallDir(packages, directory);
-    await installDir.create(recursive: true);
-
+  for (final package in packages) {
     await copyDirectoryContents(
-      tmpInstallDir.path,
+      (package as StandardPackage).sharedLibrariesDir,
       installDir.path,
       filter: (entity) => !entity.path.contains('cmake'),
     );
-  } finally {
-    await tmpDir.delete(recursive: true);
   }
-}
-
-Future<void> installNativeLibrary(
-  Package package, {
-  required String installDir,
-  required String tmpDir,
-}) async {
-  logger.fine('Installing native library ${package.libraryName}');
-
-  final packageRootDir =
-      p.join(tmpDir, '${package.library.name}-${package.version}');
-  final targetLibDir = p.join(packageRootDir, package.librariesDir);
-
-  final archiveData = await downloadUrl(package.archiveUrl);
-  await unpackArchive(
-    archiveData,
-    format: package.archiveFormat,
-    outputDir: tmpDir,
-  );
-
-  // Copy contents of lib dir from archive to install dir.
-  await copyDirectoryContents(targetLibDir, installDir);
 }
 
 LibrariesConfiguration mergedNativeLibrariesConfigurations(
@@ -83,30 +68,34 @@ LibrariesConfiguration mergedNativeLibrariesConfigurations(
   final libraryDir = mergedNativeLibrariesInstallDir(packages, directory);
 
   Package packageFor(Library library) =>
-      packages.firstWhere((package) => package.library == library);
+      packages.firstWhere((package) => package.config.library == library);
 
   return LibrariesConfiguration(
     directory: libraryDir.path,
-    enterpriseEdition: packages.first.edition == Edition.enterprise,
+    enterpriseEdition:
+        (packages.first.config as DatabasePackageConfig).edition ==
+            Edition.enterprise,
     cbl: _nativeLibraryConfiguration(packageFor(Library.libcblite)),
     cblDart: _nativeLibraryConfiguration(packageFor(Library.libcblitedart)),
   );
 }
 
 LibraryConfiguration _nativeLibraryConfiguration(Package package) {
+  final libraryName = (package as StandardPackage).libraryName;
+
   if (Platform.isMacOS || Platform.isLinux) {
     return LibraryConfiguration.dynamic(
-      package.libraryName,
+      libraryName,
       // Specifying an exact version should not be necessary, but is because
       // the beta of libcblite does not distribute properly symlinked libraries
       // for macos. We need to use the libcblite.x.dylib because that is what
       // libcblitedart is linking against.
-      version: package.version.split('.').first,
+      version: package.config.version.split('.').first,
     );
   }
 
   if (Platform.isWindows) {
-    return LibraryConfiguration.dynamic(package.libraryName);
+    return LibraryConfiguration.dynamic(libraryName);
   }
 
   throw UnsupportedError('Unsupported platform.');