+++ packages/cbl_dart/lib/src/version_info.dart
@@ -1,7 +1,8 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-import 'package.dart';
+// ignore: implementation_imports
+import 'package:cbl/src/install.dart';
 
 const latestReleases = {
-  Library.libcblite: '3.1.6',
+  Library.libcblite: '3.2.0',
   Library.libcblitedart: '7.1.0',
 };