+++ packages/cbl_dart/test/install_libraries_test.dart
@@ -1,7 +1,8 @@
 import 'dart:io';
 
+import 'package:cbl/src/install.dart';
 import 'package:cbl_dart/src/install_libraries.dart';
-import 'package:cbl_dart/src/package.dart';
+import 'package:cbl_dart/src/version_info.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
@@ -9,20 +10,23 @@ void main() {
   test('install merged native libraries', () async {
     final installDir = await Directory.systemTemp.createTemp();
 
-    final packages = Library.values.map((library) => Package(
-          library: library,
-          release: Package.latestReleases[library]!,
-          edition: Edition.enterprise,
-          target: Target.host,
-        ));
+    final loader = RemotePackageLoader();
+    final packageConfigs = DatabasePackageConfig.all(
+      releases: latestReleases,
+      edition: Edition.enterprise,
+    ).where((config) => config.os == OS.current);
+    final packages = await Future.wait(packageConfigs.map(loader.load));
 
-    await installMergedNativeLibraries(packages, directory: installDir.path);
+    await installMergedNativeLibraries(
+      packages,
+      directory: installDir.path,
+    );
 
     final installDirEntries = installDir.listSync();
     expect(installDirEntries, hasLength(1));
 
     final libDir = installDirEntries.first as Directory;
-    expect(p.basename(libDir.path), Package.mergedSignature(packages));
+    expect(p.basename(libDir.path), PackageMerging.signature(packages));
 
     final libDirEntries = libDir.listSync();
     final libDirBasenames =