+++ packages/cbl_flutter/lib/cbl_flutter.dart
@@ -3,9 +3,11 @@
 import 'dart:io';
 
 import 'package:cbl/cbl.dart';
+import 'package:cbl/src/bindings.dart';
 import 'package:cbl/src/support/isolate.dart';
 import 'package:cbl/src/support/tracing.dart';
 import 'package:cbl_flutter_platform_interface/cbl_flutter_platform_interface.dart';
+import 'package:flutter/services.dart';
 import 'package:path_provider/path_provider.dart';
 
 // ignore: avoid_classes_with_only_static_members
@@ -15,6 +17,10 @@ abstract final class CouchbaseLiteFlutter {
   /// Initializes the `cbl` package, for the main isolate.
   static Future<void> init() =>
       asyncOperationTracePoint(InitializeOp.new, () async {
+        if (Platform.isAndroid) {
+          await _preloadLibrariesForAndroid();
+        }
+
         await initPrimaryIsolate(IsolateContext(
           libraries: CblFlutterPlatform.instance.libraries(),
           initContext: await _context(),
@@ -23,24 +29,42 @@ abstract final class CouchbaseLiteFlutter {
 }
 
 Future<InitContext> _context() async {
-  final directories = await Future.wait([
-    getApplicationSupportDirectory(),
-    if (Platform.isAndroid)
-      getExternalStorageDirectory()
-    else
-      Future<Directory?>.value()
-  ]);
-
-  final filesDir = directories[0]!;
-
-  // For temporary files, we try to use the apps external storage directory
-  // and fallback to the application support directory, if it's not available.
-  final tempDir = directories[1] ?? filesDir;
-  final clbTempDir = Directory.fromUri(tempDir.uri.resolve('CBLTemp'));
-  await clbTempDir.create();
+  final filesDir = await getApplicationSupportDirectory();
+  final clbTempDir = Directory.fromUri(filesDir.uri.resolve('CBLTemp'));
+  await clbTempDir.create(recursive: true);
 
   return InitContext(
     filesDir: filesDir.path,
     tempDir: clbTempDir.path,
   );
+}
+
+/// Preloads the native libraries for Android and implements a workaround for
+/// loading of libraries on older Android versions.
+Future<void> _preloadLibrariesForAndroid() async {
+  final libraries = CblFlutterPlatform.instance.libraries();
+
+  try {
+    DynamicLibraries.fromConfig(libraries);
+    // ignore: avoid_catching_errors
+  } on ArgumentError {
+    // If we were not able to load the libraries from Dart FFI directly,
+    // load them using the Android Java API. After that, we can load the
+    // libraries from Dart FFI.
+    await _Native.instance.loadLibraries([
+      libraries.cbl.name!,
+      libraries.cblDart.name!,
+    ].map((name) => name.replaceFirst('lib', '')).toList());
+
+    DynamicLibraries.fromConfig(libraries);
+  }
+}
+
+class _Native {
+  static final instance = _Native();
+
+  final _methodChannel = const MethodChannel('cbl_flutter');
+
+  Future<void> loadLibraries(List<String> libraries) async =>
+      _methodChannel.invokeMethod('loadLibraries', libraries);
 }